+++ src/components/SQFormDialog/SQFormDialogInner.tsx
@@ -14,11 +14,13 @@ import {Form, useFormikContext} from 'formik';
 import {useDialog} from '@selectquotelabs/sqhooks';
 import {RoundedButton, DialogAlert} from 'scplus-shared-components';
 import SQFormButton from '../SQForm/SQFormButton';
-import type {DialogProps, GridProps} from '@material-ui/core';
+import type {DialogProps, GridProps, ButtonProps} from '@material-ui/core';
 import type {Theme} from '@material-ui/core/styles';
 import type {TransitionProps} from '@material-ui/core/transitions';
+import type {FormikContextType, FormikValues} from 'formik';
+import type {SQFormDialogTertiaryValue} from './types';
 
-interface SQFormDialogInnerProps {
+interface SQFormDialogInnerProps<Values extends FormikValues> {
   /** The secondary button text (Button located on left side of Dialog) */
   cancelButtonText?: string;
   /** The content to be rendered in the dialog body */
@@ -44,18 +46,22 @@ interface SQFormDialogInnerProps {
   shouldRequireFieldUpdates?: boolean;
   /** Title text at the top of the Dialog */
   title: string;
+  /**
+   * determine the status of the tertiary button
+   * can be one of'HIDE_BUTTON'|'NO_VALIDATION'|'IS_DISABLED'|'IS_ENABLED'|'FORM_VALIDATION_ONLY'|'IS_DISABLED_AND_FORM_VALIDATION',
+   * this will determine if the button is rendered, as well as if the button is disabled and if it uses form validation.
+   */
+  tertiaryStatus: SQFormDialogTertiaryValue;
   /** Any prop from https://material-ui.com/api/grid */
   muiGridProps?: GridProps;
   /** Determine if the secondary action button should be displayed */
   showSecondaryButton?: boolean;
-  /** Whether to show the tertiary button. (Default: false) */
-  showTertiaryButton?: boolean;
   /** The tertiary button text */
   tertiaryButtonText?: string;
-  /** Whether the tertiary button is disabled (Default: false) */
-  isTertiaryDisabled?: boolean;
   /** Callback function invoked when the user clicks the tertiary button */
-  onTertiaryClick?: React.MouseEventHandler<HTMLButtonElement>;
+  onTertiaryClick?: (formikContext: FormikContextType<Values>) => void;
+  /** Variant to be used for the tertiary button. Defaults to 'outlined' */
+  tertiaryButtonVariant?: ButtonProps['variant'];
 }
 
 /*
@@ -105,7 +111,7 @@ const useDialogContentStyles = makeStyles({
   },
 });
 
-function SQFormDialogInner({
+function SQFormDialogInner<Values extends FormikValues>({
   cancelButtonText,
   children,
   disableBackdropClick,
@@ -120,16 +126,27 @@ function SQFormDialogInner({
   muiGridProps,
   shouldDisplaySaveButton = true,
   showSecondaryButton = true,
-  showTertiaryButton = false,
-  isTertiaryDisabled = false,
+  tertiaryStatus = 'HIDE_BUTTON',
   onTertiaryClick,
-}: SQFormDialogInnerProps): React.ReactElement {
+  tertiaryButtonVariant,
+}: SQFormDialogInnerProps<Values>): React.ReactElement {
   const theme = useTheme();
   const titleClasses = useTitleStyles(theme);
   const actionsClasses = useActionsStyles(theme);
   const primaryActionsClasses = usePrimaryActionStyles(theme);
   const dialogContentClasses = useDialogContentStyles(theme);
-  const {resetForm, dirty: isDirty} = useFormikContext();
+  const formikContext = useFormikContext<Values>();
+
+  function getIsDisabled() {
+    switch (tertiaryStatus) {
+      case 'IS_DISABLED':
+        return true;
+      case 'FORM_VALIDATION_ONLY':
+        return !formikContext.isValid;
+      default:
+        return false;
+    }
+  }
 
   const {
     isDialogOpen: isDialogAlertOpen,
@@ -145,7 +162,7 @@ function SQFormDialogInner({
       return;
     }
 
-    if (!isDirty) {
+    if (!formikContext.dirty) {
       onClose && onClose(event, reason);
     } else {
       openDialogAlert();
@@ -153,7 +170,7 @@ function SQFormDialogInner({
   };
 
   const confirmCancel = () => {
-    resetForm();
+    formikContext.resetForm();
     onClose && onClose({}, 'escapeKeyDown');
     closeDialogAlert();
   };
@@ -181,14 +198,15 @@ function SQFormDialogInner({
 
         <Grid item={true}>
           <span style={{paddingRight: '20px'}}>
-            <SQFormButton
+            <RoundedButton
               title={tertiaryButtonText}
-              isDisabled={isTertiaryDisabled}
-              onClick={onTertiaryClick}
+              isDisabled={getIsDisabled()}
+              onClick={() => onTertiaryClick?.(formikContext)}
               type="button"
+              variant={tertiaryButtonVariant}
             >
               {tertiaryButtonText}
-            </SQFormButton>
+            </RoundedButton>
           </span>
           {shouldDisplaySaveButton && (
             <SQFormButton
@@ -232,7 +250,7 @@ function SQFormDialogInner({
               showSecondaryButton ? actionsClasses : primaryActionsClasses
             }
           >
-            {showTertiaryButton
+            {tertiaryStatus !== 'HIDE_BUTTON'
               ? renderTertiaryButton()
               : showSecondaryButton && (
                   <RoundedButton
@@ -246,7 +264,7 @@ function SQFormDialogInner({
                     {cancelButtonText}
                   </RoundedButton>
                 )}
-            {!showTertiaryButton && shouldDisplaySaveButton && (
+            {tertiaryStatus === 'HIDE_BUTTON' && shouldDisplaySaveButton && (
               <SQFormButton
                 title={saveButtonText}
                 isDisabled={isDisabled}